+++ src/Page/GroupPage/Comment/CommentList.tsx
@@ -1,206 +1,67 @@
 import styled from 'styled-components';
 import { useState } from 'react';
-import { CommentType, ReplyType } from '../../../API/Boards';
-import { GetUser } from '../GroupPageBarrel';
-import { deleteComment, deleteReply } from '../../../API/Comment';
-import DeleteModal from '../../../Component/DeleteModal';
-import { useRecoilState } from 'recoil';
-import { refreshState } from '../../../data/initialData';
-interface CommentBoxMakerProps {
-  commentData: CommentType;
-  userInfo: GetUser;
-  setInputToggle: React.Dispatch<React.SetStateAction<boolean>>;
-  setIsComment: React.Dispatch<React.SetStateAction<number>>;
-}
+import { PostProps } from '../PostPropsType';
+import { CommentToolbar } from './CommentToolbar';
+import { CommentInput } from './CommentInput';
+import CommentBoxMaker from './CommentBoxMaker';
+import Bind from '../../../Type/Bind';
 
-/** 댓글 box */
-const CommentBoxMaker = ({
-  commentData,
-  userInfo,
-  setInputToggle,
-  setIsComment,
-}: CommentBoxMakerProps) => {
-  const { profileImage, commentId } = commentData;
-  const [modalBtn, setModalBtn] = useState(false);
-  const modalBind = { state: modalBtn, Setter: setModalBtn };
-  const [refresh, setRefresh] = useRecoilState<number>(refreshState);
+const CommentList = ({ postProps }: { postProps: PostProps }): JSX.Element => {
+  const { postData, userInfo } = postProps;
 
-  // 댓글 삭제 핸들러
-  const deleteCommentHandler = async () => {
-    await deleteComment(commentId).then(() => {
-      setRefresh(refresh + 1);
-    });
-  };
+  // 댓글접기
+  const [commentFlip, setCommentFlip] = useState(true);
+  const commentFlipBind: Bind<boolean> = { state: commentFlip, Setter: setCommentFlip };
 
-  const CommentContent = () => {
-    const { nickname, createDate, content, userId } = commentData;
-    const openModal = () => setModalBtn(true);
+  // input 바텀에 붙거나 말거나
+  const [inputToggle, setInputToggle] = useState<boolean>(true);
+  const inputToggleBind: Bind<boolean> = { state: inputToggle, Setter: setInputToggle };
 
-    const addReplyHandler = () => {
-      setInputToggle(false);
-      setIsComment(commentId);
-    };
-
-    return (
-      <CommentContentS>
-        <CommentS>
-          <img src={profileImage} alt='댓글프로필' />
-          <div className='profile'>
-            <h2>{nickname}</h2>
-            <p className='date'>{createDate}</p>
-            <p className='text'>{content}</p>
-          </div>
-        </CommentS>
-        <CommentOptionS>
-          <p onClick={addReplyHandler}>답글</p>
-          {userInfo.userId === userId && (
-            <p onClick={openModal} className={'delete'}>
-              삭제
-            </p>
-          )}
-        </CommentOptionS>
-      </CommentContentS>
-    );
-  };
+  // 0이면 댓글 아니면 댓글의 commentId로 답글만들기
+  const [isComment, setIsComment] = useState<number>(0);
+  const isCommentBind: Bind<number> = { state: isComment, Setter: setIsComment };
 
   return (
     <CommentContainerS>
-      <CommentContent />
-      {commentData.replyList.map((replyData, i) => {
-        return <ReplyBoxMaker replyData={replyData} userInfo={userInfo} key={i} />;
-      })}
-      <DeleteModal modalBind={modalBind} deleteAction={deleteCommentHandler} />
+      {postData.commentCount > 0 && (
+        <>
+          <CommentToolbar postData={postData} commentFlipBind={commentFlipBind} />
+          <CommentListS commentFlip={commentFlip}>
+            {postData.commentList.map((commentData) => (
+              <CommentBoxMaker
+                userInfo={userInfo}
+                setInputToggle={setInputToggle}
+                setIsComment={setIsComment}
+                commentData={commentData}
+                key={commentData.commentId}
+              />
+            ))}
+          </CommentListS>
+        </>
+      )}
+      <CommentInput
+        postData={postData}
+        userInfo={userInfo}
+        setCommentFlip={setCommentFlip}
+        inputToggleBind={inputToggleBind}
+        isCommentBind={isCommentBind}
+      />
     </CommentContainerS>
   );
 };
 
-export default CommentBoxMaker;
-
-interface ReplyBoxMakerProps {
-  replyData: ReplyType;
-  userInfo: GetUser;
-}
-
-/** 답글 box */
-const ReplyBoxMaker = ({ replyData, userInfo }: ReplyBoxMakerProps) => {
-  const { profileImage, replyId } = replyData;
-  const [modalBtn, setModalBtn] = useState(false);
-  const modalBind = { state: modalBtn, Setter: setModalBtn };
-  const [refresh, setRefresh] = useRecoilState<number>(refreshState);
-
-  const deleteReplyHandler = async () => {
-    await deleteReply(replyId).then(() => {
-      setRefresh(refresh + 1);
-    });
-  };
-
-  const ReplyContent = (): JSX.Element => {
-    const { nickname, createDate, content, userId } = replyData;
-    const openModal = () => setModalBtn(true);
-    const WriterInfo = () => {
-      return (
-        <div>
-          <div className='profile'>
-            <h2>{nickname}</h2>
-            <p className='date'>{createDate}</p>
-          </div>
-          <p className='text'>{content}</p>
-        </div>
-      );
-    };
-
-    return (
-      <ReplyContentS>
-        <WriterInfo />
-        {userInfo.userId === userId ? (
-          <CommentOptionS>
-            <p onClick={openModal} className='delete'>
-              삭제
-            </p>
-          </CommentOptionS>
-        ) : (
-          <></>
-        )}
-      </ReplyContentS>
-    );
-  };
-
-  return (
-    <ReplyContainerS>
-      <img src={profileImage} alt='답글프로필' />
-      <ReplyContent />
-      <DeleteModal modalBind={modalBind} deleteAction={deleteReplyHandler} />
-    </ReplyContainerS>
-  );
-};
-
-const CommentS = styled.div`
-  display: flex;
-`;
+export default CommentList;
 
 const CommentContainerS = styled.div`
-  display: flex;
-  align-items: start;
-  min-height: 2rem;
-  border-radius: 1rem;
-  flex-direction: column;
-  img {
-    width: 1.875rem;
-  }
+  margin: 0 1rem;
+  margin-top: 0.5rem;
 `;
 
-const ReplyContainerS = styled(CommentContainerS)`
-  border: 2px solid var(--color-line);
-  padding: 1rem;
-`;
-
-const CommentContentS = styled.div`
-  margin-left: 0.5rem;
-
-  width: 19.0625rem;
+const CommentListS = styled.div<{ commentFlip: boolean }>`
+  height: ${(props) => (props.commentFlip ? '0px' : 'auto')};
+  margin: ${(props) => (props.commentFlip ? 'none' : '1rem 0')};
+  overflow: hidden;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
-
-  .profile {
-    display: flex;
-    align-items: center;
-    gap: var(--height-gap);
-    margin-bottom: 0.37rem;
-
-    h2 {
-      font-size: 0.875rem;
-    }
-
-    .date {
-      font-size: 0.75rem;
-      color: var(--font-color3);
-    }
-  }
-
-  p.text {
-    font-size: 0.875rem;
-    color: var(--font-color2);
-
-    p.call {
-      font-size: 0.875rem;
-      color: #000;
-      font-weight: 500;
-      display: inline;
-    }
-    margin-bottom: var(--height-gap);
-  }
-`;
-
-const ReplyContentS = styled(CommentContentS)`
-  width: 18.0625rem;
-`;
-
-const CommentOptionS = styled.div`
-  display: flex;
-  gap: 1.5rem;
-  font-size: 0.875rem;
-  .delete {
-    color: var(--font-color3);
-  }
+  gap: 0.5rem;
 `;
